+++ example/sentry.js
@@ -2,5 +2,5 @@ const { init } = require('@sentry/electron');
 
 init({
   // TODO: Replace with your project's DSN
-  dsn: '__SENTRY_DSN__',
+  dsn: '__DSN__',
 });