+++ src/index.ts
@@ -1,13 +1,28 @@
-export { ElectronOptions } from './common';
-export { ElectronFrontend } from './dispatch';
-export { captureMinidump, init, getCurrentFrontend } from './sdk';
+export {
+  Breadcrumb,
+  Request,
+  SdkInfo,
+  SentryEvent,
+  SentryException,
+  SentryResponse,
+  Severity,
+  StackFrame,
+  Stacktrace,
+  Status,
+  Thread,
+  User,
+} from '@sentry/types';
 
 export {
   addBreadcrumb,
   captureMessage,
   captureException,
   captureEvent,
-  setUserContext,
-  setTagsContext,
-  setExtraContext,
-} from '@sentry/shim';
+  configureScope,
+} from '@sentry/minimal';
+
+export { getHubFromCarrier, Hub, Scope } from '@sentry/hub';
+
+export { ElectronOptions } from './common';
+export { ElectronClient, getDefaultHub } from './dispatch';
+export { captureMinidump, init, getCurrentFrontend } from './sdk';