+++ src/renderer/backend.ts
@@ -2,26 +2,23 @@
 import { crashReporter, ipcRenderer, remote } from 'electron';
 
 import { BrowserBackend } from '@sentry/browser';
-import { Frontend, SentryError } from '@sentry/core';
-import { Breadcrumb, Context, SentryEvent } from '@sentry/shim';
+import { SentryError } from '@sentry/core';
+import { Scope } from '@sentry/hub';
+import { Breadcrumb, SentryEvent, SentryResponse } from '@sentry/types';
 
-import { CommonBackend, ElectronOptions, IPC_PING } from '../common';
+import { CommonBackend, ElectronOptions, IPC_PING, IPC_SCOPE } from '../common';
 
 /** Timeout used for registering with the main process. */
 const PING_TIMEOUT = 500;
 
 /** Backend implementation for Electron renderer backends. */
 export class RendererBackend implements CommonBackend {
-  /** Handle to the SDK frontend for callbacks. */
-  private readonly frontend: Frontend<ElectronOptions>;
-
   /** The inner SDK used to record JavaScript events. */
   private readonly inner: BrowserBackend;
 
   /** Creates a new Electron backend instance. */
-  public constructor(frontend: Frontend<ElectronOptions>) {
-    this.frontend = frontend;
-    this.inner = new BrowserBackend(frontend);
+  public constructor(private readonly options: ElectronOptions) {
+    this.inner = new BrowserBackend(options);
   }
 
   /**
@@ -59,7 +56,7 @@ export class RendererBackend implements CommonBackend {
   /**
    * @inheritDoc
    */
-  public async sendEvent(_: SentryEvent): Promise<number> {
+  public async sendEvent(_: SentryEvent): Promise<SentryResponse> {
     throw new SentryError(
       'Invariant violation: Only supported in main process',
     );
@@ -77,15 +74,13 @@ export class RendererBackend implements CommonBackend {
   /**
    * @inheritDoc
    */
-  public storeContext(_: Context): boolean {
-    throw new SentryError(
-      'Invariant violation: Only supported in main process',
-    );
+  public storeScope(scope: Scope): void {
+    ipcRenderer.send(IPC_SCOPE, scope);
   }
 
   /** Returns whether JS is enabled. */
   private isJavaScriptEnabled(): boolean {
-    return this.frontend.getOptions().enableJavaScript !== false;
+    return this.options.enableJavaScript !== false;
   }
 
   /** Returns whether native reports are enabled. */
@@ -106,7 +101,7 @@ export class RendererBackend implements CommonBackend {
       return false;
     }
 
-    return this.frontend.getOptions().enableNative !== false;
+    return this.options.enableNative !== false;
   }
 
   /** Activates the Electron CrashReporter. */
@@ -127,16 +122,20 @@ export class RendererBackend implements CommonBackend {
 
   /** Checks if the main processes is available and logs a warning if not. */
   private pingMainProcess(): void {
-    ipcRenderer.send(IPC_PING);
+    // For whatever reason we have to wait PING_TIMEOUT until we send the ping
+    // to main.
+    setTimeout(() => {
+      ipcRenderer.send(IPC_PING);
 
-    const timeout = setTimeout(() => {
-      console.warn(
-        'Could not connect to Sentry main process. Did you call init?',
-      );
+      const timeout = setTimeout(() => {
+        console.warn(
+          'Could not connect to Sentry main process. Did you call init?',
+        );
+      }, PING_TIMEOUT);
+
+      ipcRenderer.on(IPC_PING, () => {
+        clearTimeout(timeout);
+      });
     }, PING_TIMEOUT);
-
-    ipcRenderer.on(IPC_PING, () => {
-      clearTimeout(timeout);
-    });
   }
 }