+++ src/renderer/client.ts
@@ -1,20 +1,19 @@
-import { FrontendBase, Scope } from '@sentry/core';
-import { Breadcrumb, Context, SdkInfo, SentryEvent } from '@sentry/shim';
+import { BaseClient, Scope } from '@sentry/core';
+import {
+  Breadcrumb,
+  SdkInfo,
+  SentryEvent,
+  SentryResponse,
+  Status,
+} from '@sentry/types';
 // tslint:disable-next-line:no-implicit-dependencies
 import { ipcRenderer } from 'electron';
-import {
-  CommonFrontend,
-  ElectronOptions,
-  IPC_CONTEXT,
-  IPC_CRUMB,
-  IPC_EVENT,
-} from '../common';
+import { CommonClient, ElectronOptions, IPC_CRUMB, IPC_EVENT } from '../common';
 import { RendererBackend } from './backend';
 
 /** Frontend implementation for Electron renderer backends. */
-export class RendererFrontend
-  extends FrontendBase<RendererBackend, ElectronOptions>
-  implements CommonFrontend {
+export class RendererClient extends BaseClient<RendererBackend, ElectronOptions>
+  implements CommonClient {
   /**
    * Creates a new Electron SDK instance.
    * @param options Configuration options for this SDK.
@@ -26,18 +25,8 @@ export class RendererFrontend
   /**
    * @inheritDoc
    */
-  protected getSdkInfo(): SdkInfo {
+  public getSdkInfo(): SdkInfo {
     return {};
-  }
-
-  /**
-   * @inheritDoc
-   */
-  public getInitialScope(): Scope {
-    return {
-      breadcrumbs: [],
-      context: {},
-    };
   }
 
   /**
@@ -50,7 +39,7 @@ export class RendererFrontend
   public async captureMinidump(
     _path: string,
     _event: SentryEvent,
-    _scope: Scope,
+    _scope?: Scope,
   ): Promise<void> {
     // Noop
   }
@@ -58,8 +47,13 @@ export class RendererFrontend
   /**
    * @inheritDoc
    */
-  public async captureEvent(event: SentryEvent, scope: Scope): Promise<void> {
+  public async captureEvent(
+    event: SentryEvent,
+    scope?: Scope,
+  ): Promise<SentryResponse> {
     ipcRenderer.send(IPC_EVENT, event, scope);
+    // This is a fire and forget thing
+    return { code: 200, event_id: event.event_id, status: Status.Success };
   }
 
   /**
@@ -67,15 +61,8 @@ export class RendererFrontend
    */
   public async addBreadcrumb(
     breadcrumb: Breadcrumb,
-    scope: Scope,
+    scope?: Scope,
   ): Promise<void> {
     ipcRenderer.send(IPC_CRUMB, breadcrumb, scope);
-  }
-
-  /**
-   * @inheritDoc
-   */
-  public async setContext(nextContext: Context, scope: Scope): Promise<void> {
-    ipcRenderer.send(IPC_CONTEXT, nextContext, scope);
   }
 }