+++ src/renderer/index.ts
@@ -1,2 +1,14 @@
+import { defaultIntegrations } from '@sentry/browser';
+import { initAndBind } from '@sentry/core';
+import { ElectronOptions } from '..';
+import { RendererClient } from './client';
 export { RendererBackend } from './backend';
-export { RendererFrontend } from './frontend';
+export { RendererClient } from './client';
+
+/**
+ * Call init on @sentry/browser with all browser integrations
+ * @param options ElectronOptions
+ */
+export function init(options: ElectronOptions): void {
+  initAndBind(RendererClient, options, defaultIntegrations);
+}