+++ test/e2e/test-app/fixtures/sentry-onfatal-dont-exit.js
@@ -4,6 +4,6 @@ const { app } = require('electron');
 init({
   dsn: process.env.DSN,
   onFatalError: error => {
-
+    // Do nothing
   },
 });